implement Foss build flavour
Provide two build flavours 'Foss' and 'Play', the 'Play' flavour is
unchanged to what we had before. The 'Foss' flavour removes the
dependence on proprietary libraries.

This removes crashlytics, instead logging to the local log.

It also removes firebase/gcm messaging, disabling push functionality.
For now there is no replacement for this.
