+++ app/src/foss/java/chat/rocket/android/push/FirebaseTokenService.kt
@@ -1,0 +1,5 @@
+package chat.rocket.android.push
+
+class FirebaseTokenService {
+
+}
