+++ app/src/foss/java/chat/rocket/android/util/CrashlyticsManager.kt
@@ -1,0 +1,7 @@
+package chat.rocket.android.util
+
+import android.content.Context
+
+fun setupCrashlytics(context: Context) {
+        //Do absolutely nothing
+}
