+++ app/src/foss/java/chat/rocket/android/util/GMSManager.kt
@@ -1,0 +1,13 @@
+package chat.rocket.android.util
+
+
+import chat.rocket.android.main.ui.MainActivity
+import chat.rocket.android.main.presentation.MainPresenter
+
+fun refreshFCMToken(presenter: MainPresenter) {
+    //Do absolutely nothing
+}
+
+fun invalidateFirebaseToken(token: String) {
+    //Do absolutely nothing
+}
