+++ app/src/main/AndroidManifest.xml
@@ -96,23 +96,6 @@
         </receiver>
 
         <service
-            android:name=".push.FirebaseTokenService"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.google.firebase.INSTANCE_ID_EVENT" />
-            </intent-filter>
-        </service>
-
-        <service
-            android:name=".push.FirebaseMessagingService"
-            android:enabled="true"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="com.google.firebase.MESSAGING_EVENT" />
-            </intent-filter>
-        </service>
-
-        <service
             android:name=".chatroom.service.MessageService"
             android:exported="true"
             android:permission="android.permission.BIND_JOB_SERVICE" />
