+++ app/src/main/java/chat/rocket/android/app/RocketChatApplication.kt
@@ -13,15 +13,13 @@ import chat.rocket.android.dagger.DaggerAppComponent
 import chat.rocket.android.dagger.qualifier.ForMessages
 import chat.rocket.android.helper.CrashlyticsTree
 import chat.rocket.android.infrastructure.LocalRepository
-import chat.rocket.android.infrastructure.installCrashlyticsWrapper
 import chat.rocket.android.server.domain.AccountsRepository
 import chat.rocket.android.server.domain.GetCurrentServerInteractor
 import chat.rocket.android.server.domain.GetSettingsInteractor
 import chat.rocket.android.server.domain.SITE_URL
 import chat.rocket.android.server.domain.TokenRepository
 import chat.rocket.android.emoji.EmojiRepository
-import com.crashlytics.android.Crashlytics
-import com.crashlytics.android.core.CrashlyticsCore
+import chat.rocket.android.util.setupCrashlytics
 import com.facebook.drawee.backends.pipeline.DraweeConfig
 import com.facebook.drawee.backends.pipeline.Fresco
 import com.facebook.imagepipeline.core.ImagePipelineConfig
@@ -31,7 +29,6 @@ import dagger.android.DispatchingAndroidInjector
 import dagger.android.HasActivityInjector
 import dagger.android.HasBroadcastReceiverInjector
 import dagger.android.HasServiceInjector
-import io.fabric.sdk.android.Fabric
 import timber.log.Timber
 import java.lang.ref.WeakReference
 import javax.inject.Inject
@@ -89,7 +86,7 @@ class RocketChatApplication : Application(), HasActivityInjector, HasServiceInje
         AndroidThreeTen.init(this)
         EmojiRepository.load(this)
 
-        setupCrashlytics()
+        setupCrashlytics(this)
         setupFresco()
         setupTimber()
 
@@ -123,15 +120,6 @@ class RocketChatApplication : Application(), HasActivityInjector, HasServiceInje
             val message = "Server $currentServer SITE_URL"
             Timber.d(IllegalStateException(message), message)
         }
-    }
-
-    private fun setupCrashlytics() {
-        val core = CrashlyticsCore.Builder().disabled(BuildConfig.DEBUG).build()
-        Fabric.with(this, Crashlytics.Builder().core(core).build())
-
-        installCrashlyticsWrapper(this@RocketChatApplication,
-                getCurrentServerInteractor, settingsInteractor,
-                accountRepository, localRepository)
     }
 
     private fun setupFresco() {
