+++ app/src/main/java/chat/rocket/android/main/ui/MainActivity.kt
@@ -25,9 +25,9 @@ import chat.rocket.android.util.extensions.fadeIn
 import chat.rocket.android.util.extensions.fadeOut
 import chat.rocket.android.util.extensions.rotateBy
 import chat.rocket.android.util.extensions.showToast
+import chat.rocket.android.util.invalidateFirebaseToken
+import chat.rocket.android.util.refreshFCMToken
 import chat.rocket.common.model.UserStatus
-import com.google.firebase.iid.FirebaseInstanceId
-import com.google.firebase.messaging.FirebaseMessaging
 import dagger.android.AndroidInjection
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
@@ -63,13 +63,7 @@ class MainActivity : AppCompatActivity(), MainView, HasActivityInjector,
         setContentView(R.layout.activity_main)
 
         launch(CommonPool) {
-            try {
-                val token = FirebaseInstanceId.getInstance().token
-                Timber.d("FCM token: $token")
-                presenter.refreshToken(token)
-            } catch (ex: Exception) {
-                Timber.d(ex, "Missing play services...")
-            }
+            refreshFCMToken(presenter)
         }
 
         chatRoomId = intent.getStringExtra(INTENT_CHAT_ROOM_ID)
@@ -211,7 +205,7 @@ class MainActivity : AppCompatActivity(), MainView, HasActivityInjector,
     }
 
     override fun invalidateToken(token: String) =
-        FirebaseInstanceId.getInstance().deleteToken(token, FirebaseMessaging.INSTANCE_ID_SCOPE)
+        invalidateFirebaseToken(token)
 
     override fun showMessage(resId: Int) = showToast(resId)
